requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
azure-core==1.40.0
azure-identity==1.25.3
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
cryptography==47.0.0
distro==1.9.0
Django==5.0.6
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
msal==1.36.0
msal-extensions==1.3.1
openai==1.109.1
pillow==12.2.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
python-decouple==3.8
qrcode==8.2
requests==2.33.1
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
